+++ src/include/ecoz2.h
@@ -32,6 +32,18 @@ int ecoz2_prd_show_file(
         int to
         );
 
+typedef void (*hmm_learn_callback_t)(char*, long double);
+
+int ecoz2_hmm_learn(
+        int N,
+        int model_type,
+        const char* sequence_filenames[],
+        int num_sequences,
+        double hmm_epsilon,
+        double val_auto,
+        int max_iterations,
+        hmm_learn_callback_t callback
+        );
 
 #ifdef __cplusplus
 }